+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Implementing the static `#.from` method & switching constructor paradigm.
 * Fixing bug related to the `attributesUpdated` event.
 * Adding more errors related to attributes' setters.
+* Adding typed edges attributes' getters/setters.
 * Rewrote internal indices.
 
 ## 0.4.2